moving dapr bindings and triggers to FunctionApp and getting rid of DaprFunctionApp
