+++ azure/functions/__init__.py
@@ -12,7 +12,6 @@ from .decorators import (FunctionApp, Function, Blueprint,
                          AsgiFunctionApp, WsgiFunctionApp,
                          ExternalHttpFunctionApp)
 from ._durable_functions import OrchestrationContext, EntityContext
-from .decorators.dapr.dapr_function_app import DaprFunctionApp, DaprBlueprint
 from .decorators.function_app import (FunctionRegister, TriggerApi,
                                       BindingApi, SettingsApi)
 from .extension import (ExtensionMeta, FunctionExtensionException,
@@ -95,9 +94,7 @@ __all__ = (
     'AuthLevel',
     'Cardinality',
     'AccessRights',
-    'HttpMethod',
-    'DaprFunctionApp',
-    'DaprBlueprint'
+    'HttpMethod'
 )
 
 __version__ = '1.18.0b2'
